Adds basic tutorial as found on sdltutorials.com
As of now everything should compile properly if the sdl and g++ are configured on your system. It's possible that homebrew isn't needed to install the required libraries as it seemed like I already had them on machine. If all goes well then it should just create an empty window when ran.